% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico.R
\name{graficos}
\alias{graficos}
\title{Gráficos para o modelo de regressão linear}
\usage{
graficos(modelo)
}
\arguments{
\item{modelo}{um objeto gerado pela função regressao}
}
\value{
dois gráficos:
\itemize{
\item \code{preditos_vs_observados} Valores preditos (y) contra valores observados (x)
\item \code{residuos} Resíduos (y) contra valores observados (x)
}
}
\description{
Gera gráficos de valores preditos contra valores observados e um gráfico de resíduos contra valores observados.
}
