% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicao.R
\name{predicao}
\alias{predicao}
\title{Predição para novos valores}
\usage{
predicao(novos_dados, beta_hat)
}
\arguments{
\item{novos_dados}{data frame com variaveis identicas às usadas para o ajuste do modelo em \code{regressao}}

\item{beta_hat}{coeficientes providenciados por modelo$beta_hat}
}
\value{
vetor de valores preditos.
}
\description{
Retorna valores preditos para novas observações, com base no modelo ajustado por \code{regressao}
}
