% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressao.R
\name{regressao}
\alias{regressao}
\title{Modelo de regressão linear}
\usage{
regressao(dados, y)
}
\arguments{
\item{dados}{data frame com apenas variáveis numéricas, incluindo a variável resposta.}

\item{y}{nome ou posicção da coluna para ser utilizada como variável resposta.}
}
\value{
Uma lista:
\itemize{
\item \code{beta_hat} coeficientes beta estimados
\item \code{preditos} valores preditos para o banco de dados
\item \code{residuos} residuos do modelo
}
}
\description{
Ajusta um modelo de regressão linear para o data frame \code{dados} e variável resposta \code{y} fornecidos.
}
