% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salarios.R
\docType{data}
\name{salarios}
\alias{salarios}
\title{Salários e dados pessoais simulados}
\format{
\subsection{salarios}{

Um data frame com 1000 linhas e 7 colunas:
\itemize{
\item \code{id} Número de identificação da pessoa
\item \code{idade} Idade da pessoa
\item \code{altura} Altura da pessoa
\item \code{peso} Peso da pessoa
\item \code{anos_estudando} Anos que a pessoa passou estudando (escola, ensino médio, faculdade, etc)
\item \code{filhos} Número de filhos
\item \code{salario} Salário
}
}
}
\source{
data-raw/salarios.R
}
\usage{
salarios
}
\description{
Dados simulados para servirem de exemplo para as funções do pacote "trabalho2"
Todas as variáveis foram simuladas aleatóriamente, com a semente 247005. Mais detalhes podem ser encontrados no arquivo "data-raw/salarios.R"
}
\keyword{datasets}
