% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trabalho2.R
\docType{package}
\name{Trabalho2}
\alias{Trabalho2-package}
\alias{Trabalho2}
\title{Trabalho2: Ferramenta para Análise de Regressão Linear}
\description{
Um pacote para aplicar e avaliar modelos de regressão linear, com dados de exemplo.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tcaio26/ME912_trab2/}
}

}
\author{
\strong{Maintainer}: Tatiana John \email{t217505@dac.unicamp.br}

Authors:
\itemize{
  \item Caio Oliveira \email{c247005@dac.unicamp.br}
  \item Eduarda Godoy \email{e236093@dac.unicamp.br}
  \item Tiago Mattos \email{t236314@dac.unicamp.br}
}

}
